Hoist pure terminal types from the warp crate into warp_terminal
Move GridType, SizeUpdate, SizeUpdateReason, BlockPadding, and the
block list item types (InlineBannerId, InlineBannerType,
InlineBannerItem, SeparatorId) into warp_terminal. Re-export them from
their old locations so call sites do not change.

Split the pure HistoryEntry data (with PersistedCommand, ShellHost, and
LinkedWorkflowData) into app/src/terminal/history/entry.rs. This
prepares the file for a later move into warp_terminal.
