+++ app/src/terminal/mod.rs
@@ -9,7 +9,9 @@ pub use history::{
 };
 pub use view::{Event, TerminalView};
 pub use warp_terminal::shell::{self, ShellLaunchData};
-pub use warp_terminal::{CellSizeAndWindowPadding, ClipboardType, SizeInfo};
+pub use warp_terminal::{
+    BlockPadding, CellSizeAndWindowPadding, ClipboardType, SizeInfo, SizeUpdate, SizeUpdateReason,
+};
 use warpui::geometry::vector::Vector2F;
 use warpui::units::Lines;
 use warpui::{AppContext, SingletonEntity, WindowId};
@@ -176,159 +178,5 @@ pub(crate) fn element_size_at_last_frame(
         .map(|position| position.size())
 }
 
-/// The reason that the terminal size is being updated
-#[derive(Debug, Copy, Clone)]
-pub enum SizeUpdateReason {
-    /// Updated because of some general refresh (e.g. a font-size change)
-    Refresh,
-
-    /// Updated after the temrinal has been laid out, so some of the element
-    /// sizes that drive terminal size may have changed.
-    AfterLayout,
-
-    /// The shared session sharer's size changed.
-    /// This is only applicable for shared session viewers.
-    ///
-    /// The resultant [`SizeUpdate`] will use the larger of the
-    /// sharer's and viewer's size.
-    SharerSizeChanged { num_rows: usize, num_cols: usize },
-
-    /// A viewer reported its terminal size to the sharer.
-    /// This is only applicable for shared session sharers.
-    ///
-    /// The resultant [`SizeUpdate`] will use the viewer's reported
-    /// size directly (floored at 1 row and 1 column).
-    ViewerSizeReported { num_rows: usize, num_cols: usize },
-}
-
-/// Encapsulates info for updating the size of the terminal.
-#[derive(Debug, Copy, Clone)]
-pub struct SizeUpdate {
-    /// The reason for the update.
-    update_reason: SizeUpdateReason,
-
-    /// The last size info.
-    last_size: SizeInfo,
-
-    /// The new size info.
-    new_size: SizeInfo,
-
-    /// The new gap height, if there is one.
-    new_gap_height: Option<Lines>,
-
-    /// The pane-computed rows before any shared session size adjustments.
-    natural_rows: usize,
-
-    /// The pane-computed columns before any shared session size adjustments.
-    natural_cols: usize,
-}
-
-impl SizeUpdate {
-    /// Creates a size update for a layout measured directly in terminal cells.
-    pub fn from_cell_dimensions(last_size: SizeInfo, rows: usize, columns: usize) -> Self {
-        let new_size = SizeInfo::new_without_font_metrics(rows, columns);
-        Self {
-            update_reason: SizeUpdateReason::AfterLayout,
-            last_size,
-            new_size,
-            new_gap_height: None,
-            natural_rows: new_size.rows(),
-            natural_cols: new_size.columns(),
-        }
-    }
-
-    /// Returns the resulting terminal size.
-    pub fn new_size(&self) -> SizeInfo {
-        self.new_size
-    }
-
-    /// Whether the reason for the update is a refresh.
-    pub fn is_refresh(&self) -> bool {
-        matches!(self.update_reason, SizeUpdateReason::Refresh)
-    }
-
-    /// Returns whether there was any change with this update.
-    pub fn anything_changed(&self) -> bool {
-        self.pane_size_changed() || self.gap_height_changed() || self.rows_or_columns_changed()
-    }
-
-    pub fn rows_or_columns_changed(&self) -> bool {
-        self.last_size.columns() != self.new_size.columns()
-            || self.last_size.rows() != self.new_size.rows()
-    }
-
-    /// Returns whether the pane size changed with this update
-    pub fn pane_size_changed(&self) -> bool {
-        // It's fine for this to be a near-exact comparison because pane size
-        // is not determined by summing floats like content element size is.
-        (self.last_size.pane_size_px().x() - self.new_size.pane_size_px().x()).abs() > f32::EPSILON
-            || (self.last_size.pane_size_px().y() - self.new_size.pane_size_px().y()).abs()
-                > f32::EPSILON
-    }
-
-    /// Returns any new gap height to set with this update
-    pub fn new_gap_height(&self) -> Option<Lines> {
-        self.new_gap_height
-    }
-
-    /// Returns whether the gap height changed with this update
-    pub fn gap_height_changed(&self) -> bool {
-        self.new_gap_height.is_some()
-    }
-
-    /// The pane-computed natural rows before shared session adjustments.
-    pub fn natural_rows(&self) -> usize {
-        self.natural_rows
-    }
-
-    /// The pane-computed natural columns before shared session adjustments.
-    pub fn natural_cols(&self) -> usize {
-        self.natural_cols
-    }
-
-    /// Returns true if this resize was caused by a sharer size change.
-    pub fn is_sharer_size_change(&self) -> bool {
-        matches!(
-            self.update_reason,
-            SizeUpdateReason::SharerSizeChanged { .. }
-        )
-    }
-}
-
-/// The padding around each block, represented in fractional lines.
-///
-/// TODO(vorporeal): Change this to hold `Lines` instead of `f32`.
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub struct BlockPadding {
-    /// Padding from top of the block to the prompt.
-    pub padding_top: f32,
-    /// Padding from bottom of prompt to top of the command.
-    pub command_padding_top: f32,
-    /// Padding from bottom of command to top of output.
-    pub middle: f32,
-    /// Padding from bottom of output to bottom of block.
-    pub bottom: f32,
-}
-
-impl BlockPadding {
-    pub fn new(
-        padding_top: f32,
-        command_padding_top: f32,
-        padding_middle: f32,
-        padding_bottom: f32,
-    ) -> Self {
-        BlockPadding {
-            padding_top,
-            command_padding_top,
-            middle: padding_middle,
-            bottom: padding_bottom,
-        }
-    }
-}
-
 #[cfg(test)]
 mod ref_tests;
-
-#[cfg(test)]
-#[path = "size_update_tests.rs"]
-mod size_update_tests;
