+++ app/src/terminal/view.rs
@@ -139,6 +139,7 @@ use warp_core::context_flag::ContextFlag;
 use warp_core::semantic_selection::SemanticSelection;
 use warp_core::user_preferences::GetUserPreferences as _;
 use warp_errors::{report_error, report_if_error};
+pub use warp_terminal::{InlineBannerId, InlineBannerItem, InlineBannerType, SeparatorId};
 use warp_util::local_or_remote_path::LocalOrRemotePath;
 #[cfg(feature = "local_fs")]
 use warp_util::path::LineAndColumnArg;
@@ -1000,67 +1001,6 @@ impl PromptSuggestion {
         self.static_prompt_suggestion_name.is_some()
     }
 }
-
-/// A unique identifier for an inline banner.
-pub type InlineBannerId = usize;
-
-/// Type of inline banner - determines behavior like visibility in agent view.
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Hash)]
-pub enum InlineBannerType {
-    NotificationsDiscovery,
-    NotificationsError,
-    PromptSuggestions,
-    AliasExpansion,
-    SharedSessionStart,
-    SharedSessionEnd,
-    ShellProcessTerminated,
-    OpenInWarp,
-    VimMode,
-    CodebaseIndexSpeedbump,
-    AgentModeSetup,
-    AwsBedrockLogin,
-    AwsCliNotInstalled,
-}
-
-impl InlineBannerType {
-    /// Returns whether this banner type should be visible when agent view is active.
-    /// Exhaustive match ensures new banner types must define their visibility.
-    pub fn is_visible_in_agent_view(&self) -> bool {
-        match self {
-            // Agent-related banners: visible in agent view
-            Self::PromptSuggestions
-            | Self::CodebaseIndexSpeedbump
-            | Self::AgentModeSetup
-            | Self::AwsBedrockLogin
-            | Self::AwsCliNotInstalled => true,
-            // Terminal-context banners: hidden in agent view
-            Self::NotificationsDiscovery
-            | Self::NotificationsError
-            | Self::AliasExpansion
-            | Self::SharedSessionStart
-            | Self::SharedSessionEnd
-            | Self::ShellProcessTerminated
-            | Self::OpenInWarp
-            | Self::VimMode => false,
-        }
-    }
-}
-
-/// An inline banner with its unique ID and type metadata.
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Hash)]
-pub struct InlineBannerItem {
-    pub id: InlineBannerId,
-    pub banner_type: InlineBannerType,
-}
-
-impl InlineBannerItem {
-    pub fn new(id: InlineBannerId, banner_type: InlineBannerType) -> Self {
-        Self { id, banner_type }
-    }
-}
-
-/// A unique identifier for a subshell separator.
-pub type SeparatorId = usize;
 
 #[derive(Default)]
 struct InlineBannersState {
