+++ crates/warp_terminal/src/lib.rs
@@ -1,3 +1,5 @@
+mod block_list_items;
+mod block_padding;
 pub mod bootstrap;
 pub mod event;
 pub mod event_listener;
@@ -8,11 +10,15 @@ pub mod model;
 mod runtime;
 mod shared_session;
 pub mod shell;
+mod size_update;
 #[cfg(any(test, feature = "test-util"))]
 pub mod test_util;
 pub mod util;
 pub mod writeable_pty;
 
+pub use block_list_items::{InlineBannerId, InlineBannerItem, InlineBannerType, SeparatorId};
+pub use block_padding::BlockPadding;
 pub use runtime::*;
+pub use size_update::{SizeUpdate, SizeUpdateReason};
 
 pub static ASSETS: warp_assets::Assets = warp_assets::Assets;
